refactor(tools): split tools.py into a package (prompt/parse/format)
tools.py grew to 960 lines mixing three unrelated domains: prompt
rendering, model-response parsing, and result formatting. Split it into
a danyapi/tools/ package with one module per domain; __init__ re-exports
every name so existing imports keep working unchanged.
